MonitoredQuantumCircuits

Documentation for MonitoredQuantumCircuits.

MonitoredQuantumCircuits.XX

MonitoredQuantumCircuits.YY

MonitoredQuantumCircuits.ZZ

MonitoredQuantumCircuits.EmptyCircuit

MonitoredQuantumCircuits.NishimoriCircuit

MonitoredQuantumCircuits.apply!

MonitoredQuantumCircuits.apply!

MonitoredQuantumCircuits.isClifford

MonitoredQuantumCircuits.isClifford

MonitoredQuantumCircuits.nQubits

MonitoredQuantumCircuits.qiskitRepresentation

MonitoredQuantumCircuits.XX — Type

XX() <: Operation

A singelton type representing the XX operation.

MonitoredQuantumCircuits.YY — Type

YY() <: Operation

A singelton type representing the XX operation.

MonitoredQuantumCircuits.ZZ — Type

ZZ() <: Operation

A singelton type representing the XX operation.

MonitoredQuantumCircuits.EmptyCircuit — Method

EmptyCircuit(lattice::Lattice)

Create an empty circuit on the given lattice.

MonitoredQuantumCircuits.NishimoriCircuit — Method

NishimoriCircuit(lattice::Lattice)

Create a Nishimori circuit on the given lattice, i.e. a circuit with one layer of ZZ operations on all bonds.

MonitoredQuantumCircuits.apply! — Method

apply!(circuit::Circuit, executionPosition::Integer, operation::Operation, position::Vararg{Integer})

Apply the given operation at a given execution time step at the given position(s) in the circuit. The executionPosition can be used to schedule multiple operations at the same time step. However it is important to first check if the operations are compatible with each other (as of now this will show a warning which can be muted with mute=true).

MonitoredQuantumCircuits.apply! — Method

apply!(circuit::Circuit, operation::Operation, position::Vararg{Integer})

Apply the given operation at the given position(s) in the circuit. Operations that act on more than one qubit need to have the same number of position arguments as qubits they act on.

MonitoredQuantumCircuits.isClifford — Method

isClifford(circuit::Circuit)

Check if the circuit is a Clifford circuit, i.e. only contains Clifford operations. Returns true if all operations are Clifford operations, false otherwise.

MonitoredQuantumCircuits.isClifford — Method

isClifford(operation::Operation)

Return whether the operation is a Clifford operation.

MonitoredQuantumCircuits.nQubits — Method

nQubits(operation::Operation)

Return the number of qubits the operation acts on.

MonitoredQuantumCircuits.qiskitRepresentation — Method

qiskitRepresentation(operation::Operation)

Return the qiskit representation of the operation.